# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BharatSR",
    "description": "Scientifically Defensible Deep Learning Super-Resolution for Satellite Earth Observation (SIH26142)",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/export/geotiff": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Geotiff",
        "description": "Generate and download an authoritative 4-band Float32 GeoTIFF.\nSupports either pre-loaded sample_id or an active uploaded run_id.\nCRITICAL RULE:\n- If the source dataset lacks geospatial metadata, raises HTTP 400 stating\n  'No geospatial reference available'. Never silently assigns EPSG:4326.\n- If georeferenced, preserves exact CRS, scales affine transform for 4x SR (p_out = p_in / 4).",
        "operationId": "export_geotiff_api_export_geotiff_get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sample ID to export",
              "title": "Sample Id"
            },
            "description": "Sample ID to export"
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Active run ID to export",
              "title": "Run Id"
            },
            "description": "Active run ID to export"
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Model to generate SR with",
              "default": "rcan",
              "title": "Model Id"
            },
            "description": "Model to generate SR with"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/export/report": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Report",
        "description": "Generate an analytical verification report JSON for SIH evaluation.\nReports scientifically defensible metrics without fabricated claims.\nSupports either an active run_id (priority, never reruns inference)\nor sample_id (runs inference if no run artifact is provided).",
        "operationId": "export_report_api_export_report_get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sample ID",
              "title": "Sample Id"
            },
            "description": "Sample ID"
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Active run ID to export without rerunning",
              "title": "Run Id"
            },
            "description": "Active run ID to export without rerunning"
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Model ID",
              "default": "rcan",
              "title": "Model Id"
            },
            "description": "Model ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "problem_statement": {
            "type": "string",
            "title": "Problem Statement"
          },
          "target_organization": {
            "type": "string",
            "title": "Target Organization"
          },
          "sample_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "scale_factor": {
            "type": "string",
            "title": "Scale Factor"
          },
          "input_dimension": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Input Dimension"
          },
          "output_dimension": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Output Dimension"
          },
          "latency_seconds": {
            "type": "number",
            "title": "Latency Seconds"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "uncertainty_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uncertainty Summary"
          },
          "spectral_integrity_compliance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Spectral Integrity Compliance"
          },
          "geospatial_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geospatial Metadata"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "title",
          "problem_statement",
          "target_organization",
          "model_id",
          "scale_factor",
          "input_dimension",
          "output_dimension",
          "latency_seconds",
          "metrics",
          "spectral_integrity_compliance"
        ],
        "title": "ReportResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}